requirements.txt
click>=6.3
Flask>=0.11.1
Flask-Login>=0.4.0
ImageHash>=3.1
Pillow>=2.6.0
scipy>=0.18.1
